+++ CHANGES.rst
@@ -9,6 +9,7 @@
 - Result Manager: fixed bug when working dir contained a revision with number 0 (#822)
 - Result Manager: grid and result id (uuid) are now saved in project file.
 - Result Manager: updated flow of deletion signals (layers are unloaded after grid/results are removed from tools)
+- Result Manager: fixed bug when pressing delete button with empty model.
 - Graph tool: maptools are disabled when tabs are switched (#824)
 - Graph tool: Fixed bug where deactivated plots were added again when parameter/units or absoluteness was changed (#825)
 - Graph tool: default variable for pump should be 'discharge pump' (#819)